@@ -41,7 +41,8 @@
     "@babel/preset-react": "^7.12.10",
     "@date-io/date-fns": "^1.3.13",
     "@mui/x-date-pickers": "^5.0.0-alpha.7",
-    "@testing-library/react": "^12.0.0",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
     "@types/jest": "^27.0.3",
     "@wojtekmaj/enzyme-adapter-react-17": "^0.6.6",
     "babel-loader": "^8.2.2",
@@ -57,8 +58,8 @@
     "jest": "^27.3.1",
     "prettier": "2.3.2",
     "prop-types": "^15.7.2",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "source-map-loader": "^4.0.1",
     "ts-jest": "^27.0.7",
     "webpack": "^5.17.0",
